backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
gunicorn==23.0.0
python-dotenv==1.0.1
motor==3.6.0
pymongo==4.9.0
pydantic>=2.11.5
email-validator==2.2.0
bcrypt==4.2.0
PyJWT==2.9.0
mistralai>=2.0.0
python-docx==1.1.2
pdfminer.six==20231228
pypdf>=5.1.0
pdfplumber==0.11.4
pytesseract==0.3.13
pdf2image==1.17.0
Jinja2==3.1.4
python-multipart==0.0.9
httpx>=0.28.1
slowapi==0.1.9
pytest==8.3.3
pytest-asyncio==0.24.0
llama-index-core==0.14.13
llama-index-llms-mistralai==0.9.0
xhtml2pdf==0.2.16
llama-index-embeddings-mistralai==0.4.1
llama-index-readers-file==0.5.6
diskcache==5.6.3
